+++ rust/src/core_mechanics/damage_dice.rs
@@ -1,17 +1,17 @@
 #[derive(Clone)]
 pub struct DamageDice {
-    count: i8,
-    increments: i8,
-    size: i8,
+    count: i32,
+    increments: i32,
+    size: i32,
 }
 
 // Commonly used damage dice for weapon and spell definitions
-pub static D6: i8 = 4;
-pub static D8: i8 = 5;
-pub static D10: i8 = 6;
+pub static D6: i32 = 4;
+pub static D8: i32 = 5;
+pub static D10: i32 = 6;
 
 impl DamageDice {
-    pub fn new(increments: i8) -> DamageDice {
+    pub fn new(increments: i32) -> DamageDice {
         // 5d10+ has different scaling
         if increments >= 13 {
             return DamageDice {
@@ -22,7 +22,7 @@ impl DamageDice {
         }
 
         let mut increments_mut = increments;
-        let mut count: i8 = 1;
+        let mut count: i32 = 1;
         while increments_mut > 6 {
             increments_mut -= 3;
             count *= 2;
@@ -40,7 +40,7 @@ impl DamageDice {
         return DamageDice { count, increments, size };
     }
 
-    pub fn add(&self, increments: i8) -> DamageDice {
+    pub fn add(&self, increments: i32) -> DamageDice {
         Self::new(self.increments + increments)
     }
 
@@ -56,11 +56,11 @@ impl DamageDice {
         return ((self.count * (self.size + 1)) as f64) / 2.0
     }
 
-    pub fn aoe_damage(rank: i8) -> Self {
+    pub fn aoe_damage(rank: i32) -> Self {
         return Self::new(D8 + (rank - 1));
     }
 
-    pub fn single_target_damage(rank: i8) -> Self {
+    pub fn single_target_damage(rank: i32) -> Self {
         return Self::new(D10 + (rank - 1));
     }
 }
