+++ rust/src/latex_formatting.rs
@@ -55,7 +55,7 @@ pub fn uppercase_first_letter(text: &str) -> String {
     }
 }
 
-pub fn modifier(val: i8) -> String {
+pub fn modifier(val: i32) -> String {
     if val >= 0 {
         return format!("+{}", val)
     } else {
